+++ samples/liteuart/sim_main.cpp
@@ -42,9 +42,9 @@ void Init() {
     //=================================================
     bus->wb_clk = &top->clk100;
     bus->wb_rst = &top->rst0;
-    bus->wb_addr = (unsigned long long *)&top->wb_adr;
-    bus->wb_rd_dat = (unsigned long long *)&top->wb_dat_r;
-    bus->wb_wr_dat = (unsigned long long *)&top->wb_dat_w;
+    bus->wb_addr = (uint64_t *)&top->wb_adr;
+    bus->wb_rd_dat = (uint64_t *)&top->wb_dat_r;
+    bus->wb_wr_dat = (uint64_t *)&top->wb_dat_w;
     bus->wb_we = &top->wb_we;
     bus->wb_sel = &top->wb_sel;
     bus->wb_stb = &top->wb_stb;
